@@ -1,119 +1,118 @@
 {
   "name": "xy-diary",
-  "version": "0.1.0",
-  "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "serve": "npm run renewal && npm run dev:h5",
-    "serve:wx": "npm run renewal && npm run dev:mp-weixin",
-    "build": "VUE_APP_ENV=online npm run renewal && npm run build:h5",
-    "build:wx": "VUE_APP_ENV=online npm run renewal && npm run build:mp-weixin",
-    "build:toutiao": "VUE_APP_ENV=online npm run renewal && npm run build:mp-toutiao",
-    "lint": "vue-cli-service lint",
-    "renewal": "node src/newPages.js && node src/manifest.js",
-    "build:app-plus": "cross-env NODE_ENV=production UNI_PLATFORM=app-plus vue-cli-service uni-build",
-    "build:custom": "cross-env NODE_ENV=production uniapp-cli custom",
-    "build:h5": "cross-env VUE_APP_ENV=online NODE_ENV=production UNI_PLATFORM=h5 vue-cli-service uni-build",
-    "build:h5test": "cross-env VUE_APP_ENV=qa NODE_ENV=production UNI_PLATFORM=h5 vue-cli-service uni-build",
-    "build:mp-360": "cross-env NODE_ENV=production UNI_PLATFORM=mp-360 vue-cli-service uni-build",
-    "build:mp-alipay": "cross-env NODE_ENV=production UNI_PLATFORM=mp-alipay vue-cli-service uni-build",
-    "build:mp-baidu": "cross-env NODE_ENV=production UNI_PLATFORM=mp-baidu vue-cli-service uni-build",
-    "build:mp-kuaishou": "cross-env NODE_ENV=production UNI_PLATFORM=mp-kuaishou vue-cli-service uni-build",
-    "build:mp-qq": "cross-env NODE_ENV=production UNI_PLATFORM=mp-qq vue-cli-service uni-build",
-    "build:mp-toutiao": "cross-env NODE_ENV=production UNI_PLATFORM=mp-toutiao vue-cli-service uni-build",
-    "build:mp-weixin": "cross-env NODE_ENV=production UNI_PLATFORM=mp-weixin vue-cli-service uni-build",
-    "build:qa": "VUE_APP_ENV=qa npm run renewal && npm run build:h5test",
-    "build:quickapp-native": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-native vue-cli-service uni-build",
-    "build:quickapp-webview": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview vue-cli-service uni-build",
-    "build:quickapp-webview-huawei": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview-huawei vue-cli-service uni-build",
-    "build:quickapp-webview-union": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview-union vue-cli-service uni-build",
-    "dev:app-plus": "cross-env NODE_ENV=development UNI_PLATFORM=app-plus vue-cli-service uni-build --watch",
-    "dev:custom": "cross-env NODE_ENV=development VUE_APP_ENV=development uniapp-cli custom",
-    "dev:h5": "cross-env VUE_APP_ENV=dev NODE_ENV=development UNI_PLATFORM=h5 vue-cli-service uni-serve",
-    "dev:mp-360": "cross-env NODE_ENV=development UNI_PLATFORM=mp-360 vue-cli-service uni-build --watch",
-    "dev:mp-alipay": "cross-env NODE_ENV=development UNI_PLATFORM=mp-alipay vue-cli-service uni-build --watch",
-    "dev:mp-baidu": "cross-env NODE_ENV=development UNI_PLATFORM=mp-baidu vue-cli-service uni-build --watch",
-    "dev:mp-kuaishou": "cross-env NODE_ENV=development UNI_PLATFORM=mp-kuaishou vue-cli-service uni-build --watch",
-    "dev:mp-qq": "cross-env NODE_ENV=development UNI_PLATFORM=mp-qq vue-cli-service uni-build --watch",
-    "dev:mp-toutiao": "cross-env NODE_ENV=development UNI_PLATFORM=mp-toutiao vue-cli-service uni-build --watch",
-    "dev:mp-weixin": "cross-env NODE_ENV=development UNI_PLATFORM=mp-weixin vue-cli-service uni-build --watch",
-    "dev:quickapp-native": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-native vue-cli-service uni-build --watch",
-    "dev:quickapp-webview": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview vue-cli-service uni-build --watch",
-    "dev:quickapp-webview-huawei": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview-huawei vue-cli-service uni-build --watch",
-    "dev:quickapp-webview-union": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview-union vue-cli-service uni-build --watch",
-    "info": "node node_modules/@dcloudio/vue-cli-plugin-uni/commands/info.js",
-    "serve:quickapp-native": "node node_modules/@dcloudio/uni-quickapp-native/bin/serve.js",
-    "test:android": "cross-env UNI_PLATFORM=app-plus UNI_OS_NAME=android jest -i",
-    "test:h5": "cross-env UNI_PLATFORM=h5 jest -i",
-    "test:ios": "cross-env UNI_PLATFORM=app-plus UNI_OS_NAME=ios jest -i",
-    "test:mp-baidu": "cross-env UNI_PLATFORM=mp-baidu jest -i",
-    "test:mp-weixin": "cross-env UNI_PLATFORM=mp-weixin jest -i"
+    "dev:app-android": "uni -p app-android",
+    "dev:app-ios": "uni -p app-ios",
+    "dev:app-harmony": "uni -p app-harmony",
+    "dev:custom": "uni -p",
+    "dev:h5": "uni",
+    "dev:h5:ssr": "uni --ssr",
+    "dev:mp-alipay": "uni -p mp-alipay",
+    "dev:mp-baidu": "uni -p mp-baidu",
+    "dev:mp-jd": "uni -p mp-jd",
+    "dev:mp-kuaishou": "uni -p mp-kuaishou",
+    "dev:mp-lark": "uni -p mp-lark",
+    "dev:mp-qq": "uni -p mp-qq",
+    "dev:mp-toutiao": "uni -p mp-toutiao",
+    "dev:mp-weixin": "uni -p mp-weixin",
+    "dev:mp-xhs": "uni -p mp-xhs",
+    "dev:quickapp-webview": "uni -p quickapp-webview",
+    "dev:quickapp-webview-huawei": "uni -p quickapp-webview-huawei",
+    "dev:quickapp-webview-union": "uni -p quickapp-webview-union",
+    "build:app": "uni build -p app",
+    "build:app-android": "uni build -p app-android",
+    "build:app-ios": "uni build -p app-ios",
+    "build:app-harmony": "uni build -p app-harmony",
+    "build:custom": "uni build -p",
+    "build:h5": "uni build",
+    "build:h5:ssr": "uni build --ssr",
+    "build:mp-alipay": "uni build -p mp-alipay",
+    "build:mp-baidu": "uni build -p mp-baidu",
+    "build:mp-jd": "uni build -p mp-jd",
+    "build:mp-kuaishou": "uni build -p mp-kuaishou",
+    "build:mp-lark": "uni build -p mp-lark",
+    "build:mp-qq": "uni build -p mp-qq",
+    "build:mp-toutiao": "uni build -p mp-toutiao",
+    "build:mp-weixin": "uni build -p mp-weixin",
+    "build:mp-xhs": "uni build -p mp-xhs",
+    "build:quickapp-webview": "uni build -p quickapp-webview",
+    "build:quickapp-webview-huawei": "uni build -p quickapp-webview-huawei",
+    "build:quickapp-webview-union": "uni build -p quickapp-webview-union",
+    "type-check": "vue-tsc --noEmit"
   },
   "dependencies": {
-    "@dcloudio/uni-app-plus": "^2.0.0-32220210818002",
-    "@dcloudio/uni-h5": "^2.0.0-32220210818002",
-    "@dcloudio/uni-helper-json": "*",
-    "@dcloudio/uni-i18n": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-360": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-alipay": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-baidu": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-kuaishou": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-qq": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-toutiao": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-vue": "^2.0.0-32220210818002",
-    "@dcloudio/uni-mp-weixin": "^2.0.0-32220210818002",
-    "@dcloudio/uni-quickapp-native": "^2.0.0-32220210818002",
-    "@dcloudio/uni-quickapp-webview": "^2.0.0-32220210818002",
-    "@dcloudio/uni-stat": "^2.0.0-32220210818002",
-    "@lucky-canvas/uni": "0.0.13",
-    "@lucky-canvas/vue": "^0.1.11",
-    "@vue/shared": "^3.0.0",
-    "core-js": "^3.6.5",
-    "howler": "^2.2.4",
-    "moment": "^2.29.4",
-    "regenerator-runtime": "^0.12.1",
-    "uview-ui": "^1.8.4",
-    "vconsole": "^3.9.1",
-    "vue": "^2.6.11",
-    "vuex": "^3.2.0"
+    "@dcloudio/uni-app": "3.0.0-4020420240722002",
+    "@dcloudio/uni-app-harmony": "3.0.0-4020420240722002",
+    "@dcloudio/uni-app-plus": "3.0.0-4020420240722002",
+    "@dcloudio/uni-components": "3.0.0-4020420240722002",
+    "@dcloudio/uni-h5": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-alipay": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-baidu": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-jd": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-kuaishou": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-lark": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-qq": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-toutiao": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-weixin": "3.0.0-4020420240722002",
+    "@dcloudio/uni-mp-xhs": "3.0.0-4020420240722002",
+    "@dcloudio/uni-quickapp-webview": "3.0.0-4020420240722002",
+    "@lucky-canvas/uni": "^0.0.13",
+    "axios": "^1.7.3",
+    "clipboard": "^2.0.11",
+    "dayjs": "^1.11.12",
+    "md5": "^2.3.0",
+    "mitt": "^3.0.1",
+    "pinia": "^2.2.1",
+    "pinia-plugin-persistedstate": "^3.2.1",
+    "uview-plus": "^3.3.16",
+    "vue": "^3.4.37",
+    "vue-i18n": "^9.13.1",
+    "vuedraggable": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/runtime": "~7.12.0",
-    "@dcloudio/types": "*",
-    "@dcloudio/uni-automator": "^2.0.0-32220210818002",
-    "@dcloudio/uni-cli-shared": "^2.0.0-32220210818002",
-    "@dcloudio/uni-migration": "^2.0.0-32220210818002",
-    "@dcloudio/uni-template-compiler": "^2.0.0-32220210818002",
-    "@dcloudio/vue-cli-plugin-hbuilderx": "^2.0.0-32220210818002",
-    "@dcloudio/vue-cli-plugin-uni": "^2.0.0-32220210818002",
-    "@dcloudio/vue-cli-plugin-uni-optimize": "^2.0.0-32220210818002",
-    "@dcloudio/webpack-uni-mp-loader": "^2.0.0-32220210818002",
-    "@dcloudio/webpack-uni-pages-loader": "^2.0.0-32220210818002",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-import": "^1.11.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "jest": "^25.4.0",
-    "mini-types": "*",
-    "miniprogram-api-typings": "*",
-    "postcss-comment": "^2.0.0",
-    "sass": "^1.38.1",
-    "sass-loader": "^10.1.1",
-    "vue-template-compiler": "^2.6.11"
+    "@dcloudio/types": "^3.4.12",
+    "@dcloudio/uni-automator": "3.0.0-4020420240722002",
+    "@dcloudio/uni-cli-shared": "3.0.0-4020420240722002",
+    "@dcloudio/uni-stacktracey": "3.0.0-4020420240722002",
+    "@dcloudio/vite-plugin-uni": "3.0.0-4020420240722002",
+    "@types/md5": "^2.3.5",
+    "@types/node": "^20.14.15",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^7.18.0",
+    "@vue/runtime-core": "^3.4.37",
+    "@vue/tsconfig": "^0.1.3",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-vue": "^9.27.0",
+    "prettier": "^3.3.3",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "sass": "^1.77.8",
+    "sass-loader": "10",
+    "stylelint": "^16.8.1",
+    "stylelint-config-html": "^1.1.0",
+    "stylelint-config-recess-order": "^5.0.1",
+    "stylelint-config-recommended-scss": "^14.1.0",
+    "stylelint-config-recommended-vue": "^1.5.0",
+    "stylelint-config-standard": "^36.0.1",
+    "stylelint-config-standard-scss": "^13.1.0",
+    "typescript": "^4.9.5",
+    "unplugin-auto-import": "^0.17.8",
+    "unplugin-vue-components": "^0.26.0",
+    "unplugin-vue-setup-extend-plus": "^1.0.1",
+    "vite": "^5.4.0",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-vue-devtools": "^7.3.7",
+    "vue-tsc": "^1.8.27"
   },
-  "browserslist": [
-    "Android >= 4",
-    "ios >= 8"
-  ],
-  "uni-app": {
-    "scripts": {}
+  "engines": {
+    "node": ">=16.18.0"
+  },
+  "config": {
+    "commitizen": {
+      "path": "node_modules/cz-git"
+    }
   }
 }
