@@ -1,39 +1,30 @@
 <template>
-  <view class="empty" :style="{height: height + 'px'}">
-    <image class="empty-img" src="@/static/image/home/5.png"></image>
-    <text class="empty-text" v-html="text"/>
-    <slot/>
+  <view :style="{ height: props.height + 'px' }" class="empty">
+    <image class="empty-img" src="https://s21.ax1x.com/2024/08/19/pAPlLUU.png"></image>
+    <text class="empty-text" v-html="props.text" />
+    <slot />
   </view>
 </template>
 
-<script>
-export default {
-  name: 'index',
-  props: {
-    height: {
-      type: [Number, String],
-      default: 500
-    },
-    text: {
-      type: String,
-      default: '暂无数据'
-    }
-  },
-  data () {
-    return {}
-  }
+<script lang="ts" setup>
+interface Props {
+  height: number | string;
+  text: string;
 }
+const props = withDefaults(defineProps<Props>(), {
+  height: 500,
+  text: "暂无数据"
+});
 </script>
 
 <style lang="scss" scoped>
-
-.empty{
+.empty {
   @include flexCenter;
   flex-direction: column;
-  &-img{
+  &-img {
     @include wh(360rpx, 360rpx);
   }
-  &-text{
+  &-text {
     color: $uni-color;
     font-size: 30rpx;
     text-align: center;
