@@ -7,16 +7,6 @@
   </view>
 </template>
 
-<script>
-export default {
-  name: 'loading3',
-  data () {
-    return {
-    }
-  }
-}
-</script>
-
 <style scoped="true">
 .container {
   width: 30px;
@@ -24,9 +14,9 @@ export default {
   position: relative;
 }
 
- .container.loading3 {
+.container.loading3 {
   -webkit-animation: rotation 1s infinite;
-          animation: rotation 1s infinite;
+  animation: rotation 1s infinite;
 }
 .container.loading3 .shape1 {
   border-top-left-radius: 10px;
@@ -49,124 +39,124 @@ export default {
 }
 .container .shape.shape1 {
   left: 0;
-  background-color: #1890FF;
+  background-color: #1890ff;
 }
 .container .shape.shape2 {
   right: 0;
-  background-color: #91CB74;
+  background-color: #91cb74;
 }
 .container .shape.shape3 {
   bottom: 0;
-  background-color: #FAC858;
+  background-color: #fac858;
 }
 .container .shape.shape4 {
   bottom: 0;
   right: 0;
-  background-color: #EE6666;
+  background-color: #ee6666;
 }
 
 .loading3 .shape1 {
   -webkit-animation: animation3shape1 0.5s ease 0s infinite alternate;
-          animation: animation3shape1 0.5s ease 0s infinite alternate;
+  animation: animation3shape1 0.5s ease 0s infinite alternate;
 }
 
 @-webkit-keyframes animation3shape1 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(5px, 5px);
-            transform: translate(5px, 5px);
+    transform: translate(5px, 5px);
   }
 }
 
 @keyframes animation3shape1 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(5px, 5px);
-            transform: translate(5px, 5px);
+    transform: translate(5px, 5px);
   }
 }
 .loading3 .shape2 {
   -webkit-animation: animation3shape2 0.5s ease 0s infinite alternate;
-          animation: animation3shape2 0.5s ease 0s infinite alternate;
+  animation: animation3shape2 0.5s ease 0s infinite alternate;
 }
 
 @-webkit-keyframes animation3shape2 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(-5px, 5px);
-            transform: translate(-5px, 5px);
+    transform: translate(-5px, 5px);
   }
 }
 
 @keyframes animation3shape2 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(-5px, 5px);
-            transform: translate(-5px, 5px);
+    transform: translate(-5px, 5px);
   }
 }
 .loading3 .shape3 {
   -webkit-animation: animation3shape3 0.5s ease 0s infinite alternate;
-          animation: animation3shape3 0.5s ease 0s infinite alternate;
+  animation: animation3shape3 0.5s ease 0s infinite alternate;
 }
 
 @-webkit-keyframes animation3shape3 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(5px, -5px);
-            transform: translate(5px, -5px);
+    transform: translate(5px, -5px);
   }
 }
 
 @keyframes animation3shape3 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(5px, -5px);
-            transform: translate(5px, -5px);
+    transform: translate(5px, -5px);
   }
 }
 .loading3 .shape4 {
   -webkit-animation: animation3shape4 0.5s ease 0s infinite alternate;
-          animation: animation3shape4 0.5s ease 0s infinite alternate;
+  animation: animation3shape4 0.5s ease 0s infinite alternate;
 }
 
 @-webkit-keyframes animation3shape4 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(-5px, -5px);
-            transform: translate(-5px, -5px);
+    transform: translate(-5px, -5px);
   }
 }
 
 @keyframes animation3shape4 {
   from {
     -webkit-transform: translate(0, 0);
-            transform: translate(0, 0);
+    transform: translate(0, 0);
   }
   to {
     -webkit-transform: translate(-5px, -5px);
-            transform: translate(-5px, -5px);
+    transform: translate(-5px, -5px);
   }
 }
 </style>
